Changes File Structure
