+++ package.json
@@ -9,6 +9,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "axios": "^1.1.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
